add shared assertions
