+++ mcache/tests/test_integration_e2e.py
@@ -8,6 +8,37 @@ from .metrics import GAUGES, ITEMS_GAUGES, ITEMS_RATES, RATES, SLABS_AGGREGATES,
 from .utils import count_connections, get_host_socket_path
 
 
+def assert_check_coverage(aggregator):
+    """
+    Check coverage
+    """
+    expected_tags = ["url:{}:{}".format(HOST, PORT), 'foo:bar']
+    for m in GAUGES + RATES + SLABS_AGGREGATES:
+        aggregator.assert_metric(m, tags=expected_tags)
+
+    expected_tags += ["slab:1"]
+    for m in ITEMS_GAUGES + ITEMS_RATES + SLABS_RATES + SLABS_GAUGES:
+        aggregator.assert_metric(m, tags=expected_tags)
+
+    aggregator.assert_all_metrics_covered()
+
+
+@pytest.mark.e2e
+def test_e2e(client, dd_agent_check, instance):
+    """
+    Test all the available metrics: default, options and slabs
+    """
+    # we need to successfully retrieve a key to produce `get_hit_percent`
+    for _ in range(100):
+        assert client.set("foo", "bar") is True
+        assert client.get("foo") == "bar"
+
+    instance.update({'options': {'items': True, 'slabs': True}})
+    aggregator = dd_agent_check(instance, rate=True)
+
+    assert_check_coverage(aggregator)
+
+
 @pytest.mark.integration
 @pytest.mark.usefixtures('dd_environment')
 def test_service_ok(check, instance, aggregator):
@@ -36,15 +67,7 @@ def test_metrics(client, check, instance, aggregator):
     instance.update({'options': {'items': True, 'slabs': True}})
     check.check(instance)
 
-    expected_tags = ["url:{}:{}".format(HOST, PORT), 'foo:bar']
-    for m in GAUGES + RATES + SLABS_AGGREGATES:
-        aggregator.assert_metric(m, tags=expected_tags, count=1)
-
-    expected_tags += ["slab:1"]
-    for m in ITEMS_GAUGES + ITEMS_RATES + SLABS_RATES + SLABS_GAUGES:
-        aggregator.assert_metric(m, tags=expected_tags, count=1)
-
-    assert aggregator.metrics_asserted_pct == 100.0
+    assert_check_coverage(aggregator)
 
 
 @requires_socket_support
